prod/requirements.txt
streamlit>=1.33
torch
torchvision
pillow